Rename regexText to utils & Move into src (#33)
